feat:修改mybatis生成器

+++ src/main/java/com/mida/chromeext/mapper/BgPictureMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface BgPictureMapper {
-    int countByExample(BgPictureExample example);
+    long countByExample(BgPictureExample example);
 
     int deleteByExample(BgPictureExample example);
