+++ src/main/java/com/mida/chromeext/mapper/SiteMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface SiteMapper {
-    int countByExample(SiteExample example);
+    long countByExample(SiteExample example);
 
     int deleteByExample(SiteExample example);
